main.cpp: fix bug: u8 is between 0-255
FIXME: if user doesn't press select, LoadPerGameSettings doesn't load
and notification led keep turn off

+++ gui/source/settings.h
@@ -94,9 +94,9 @@ typedef struct _Settings_t {
 		s8 extvram;
 		s8 lockarm9scfgext;
 		s8 donor;
-		u8 red;
-		u8 green;
-		u8 blue;
+		int red;
+		int green;
+		int blue;
 	} pergame;
 } Settings_t;
 extern Settings_t settings;
